169 - Relocate BranchFilter and TestIcon to utils folder
