+++ frontend/src/components/card/SeriesCard.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { pickIcon } from '../TestIcon';
+import { pickIcon } from '../../utils/TestIcon';
 import { useTranslation } from 'react-i18next';
 import { CardValue, CardSubTitle, CardTitle } from './SeriesCard.styles';
 import {
