+++ frontend/src/components/historyTable/TestStatus.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useStateValue } from '../../contexts/state';
-import { pickIcon } from '../TestIcon';
+import { pickIcon } from '../../utils/TestIcon';
 import { DefinedData } from './TestStatus.styles';
 import { addBgColor, removeBgColor } from './Heading';
