+++ frontend/src/components/suite/Testlist.jsx
@@ -1,6 +1,6 @@
 ﻿import React, { useState } from 'react';
 import { useParams } from 'react-router';
-import { pickIcon } from '../TestIcon';
+import { pickIcon } from '../../utils/TestIcon';
 import {
     FlexContainer,
     HeaderContainer,
